pyproject.toml
[project]
name = "patstat-mcp-lite"
version = "0.1.0"
description = "MCP server for PATSTAT schema discovery & query execution on EPO TIP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.8.0,<1.26.0",
    "starlette>=0.27.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "pandas>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/patstat_mcp"]

[project.scripts]
patstat-mcp-lite = "patstat_mcp.server:main"
